Removed the dotnet project files.
I spent a lot of time trying to get this to work, but the quality of the template + it's documentation is utter crap. If I had spent this much time starting a node api from scratch, I would have been done in half the time that I wasted on this so far.